test : 테스트 코드 BlackPawn, WhitePawn 분리
