+++ src/test/java/chess/domain/piece/PawnTest.java
@@ -9,7 +9,6 @@ import org.junit.jupiter.api.Test;
 import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class PawnTest {
 
@@ -27,85 +26,6 @@ class PawnTest {
     public static final Position D5 = new Position(File.D, Rank.FIVE);
 
     @Test
-    @DisplayName("흑색 폰의 대각선 경로를 계산한다.")
-    void computePath_Black() {
-        final var pawn = new BlackPawn();
-
-        assertThat(pawn.computePathWithValidate(B5, A4)).containsExactly(A4);
-    }
-
-    @Test
-    @DisplayName("백색 폰의 한칸 전진 경로를 계산한다.")
-    void computePath_White() {
-        final var pawn = new WhitePawn();
-
-        assertThat(pawn.computePathWithValidate(B5, B6)).containsExactly(B6);
-    }
-
-    @Test
-    @DisplayName("백색 폰의 두칸 전진 경로를 계산한다.")
-    void computePath_WhiteTwo() {
-        final var pawn = new WhitePawn();
-
-        assertThat(pawn.computePathWithValidate(B2, B4)).containsExactlyInAnyOrder(B3, B4);
-    }
-
-    @Test
-    @DisplayName("폰이 시작위치가 아니면 두칸 전진할 수 없다.")
-    void computePath_WhiteTwoException() {
-        final var pawn = new WhitePawn();
-
-        assertThatThrownBy(() -> pawn.computePathWithValidate(B5, B7))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    @DisplayName("흑색 폰의 두칸 전진 경로를 계산한다.")
-    void computePath_BlackTwo() {
-        final var pawn = new BlackPawn();
-
-        assertThat(pawn.computePathWithValidate(B7, B5)).containsExactlyInAnyOrder(B6, B5);
-    }
-
-    @Test
-    @DisplayName("폰이 초기위치가 아니면 두칸 전진할 수 없다.")
-    void computePath_BlackTwoException() {
-        final var pawn = new BlackPawn();
-
-        assertThatThrownBy(() -> pawn.computePathWithValidate(B5, B3))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    @DisplayName("유효하지 않은 위치를 받으면 예외가 발생한다.")
-    void computePath_BlackIllegalTarget_Exception() {
-        final var pawn = new BlackPawn();
-
-        assertThatThrownBy(() -> pawn.computePathWithValidate(B5, E4))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    @DisplayName("유효하지 않은 위치를 받으면 예외가 발생한다.")
-    void computePath_WhiteIllegalTarget_Exception() {
-        final var pawn = new WhitePawn();
-
-        assertThatThrownBy(() -> pawn.computePathWithValidate(B5, E4))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    @DisplayName("폰은 대각선 위치에 기물이 있으면 이동할 수 있다.")
-    void canMove_diagonalExist_true() {
-        final var pawn = new WhitePawn();
-
-        final var source = C4;
-        final var target = D5;
-
-        assertThat(pawn.canMove(Map.of(target, false), source, target)).isTrue();
-    }
-
-    @Test
     @DisplayName("폰은 전진할 경로에 기물이 없으면 이동할 수 있다.")
     void canMove_forwardEmpty_true() {
         final var pawn = new WhitePawn();
@@ -114,28 +34,6 @@ class PawnTest {
         final var target = C5;
 
         assertThat(pawn.canMove(Map.of(target, true), source, target)).isTrue();
-    }
-
-    @Test
-    @DisplayName("백색 폰은 아래로 갈 수 없다.")
-    void canMove_down_false() {
-        final var pawn = new WhitePawn();
-
-        final var source = C5;
-        final var target = C4;
-
-        assertThatThrownBy(() -> pawn.computePathWithValidate(source, target)).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @Test
-    @DisplayName("흑색 폰은 위로 갈 수 없다.")
-    void canMove_up_false() {
-        final var pawn = new BlackPawn();
-
-        final var source = C4;
-        final var target = C5;
-
-        assertThatThrownBy(() -> pawn.computePathWithValidate(source, target)).isInstanceOf(IllegalArgumentException.class);
     }
 
     @Test
